[FEATURE] Add work counter
A singular function is in charge of updating the
work counter and its associated label.

When the work timer finishes, this counter is
decremented; when it reaches zero, the long break
timer is automatically selected. Otherwise, the
short break timer is selected instead.

Aside from that, the following was done:
* Make some of the widgets unique.
* Randomize alarm pitch slightly
* Unpause timer in _stop_timer()